Added new animations. Modified mode B for moving button animations. Changed ultra mode for mode b.
todo:
Actual animation for the buttons

+++ Assets/Scripts/LivingBody.cs
@@ -20,7 +20,7 @@ public class LivingBody : MonoBehaviour {
         Start();
     }
 
-    public virtual void TakeHit(int hitAmount = 1)
+    public virtual void TakeHit(float hitAmount = 1)
     {
       
         currentHealth -= hitAmount;
